requirements.txt
setuptools == 59.5.0
torchmetrics~=0.11.3
torchsummary == 1.5.1
jsonargparse[signatures]~=4.20.0 # -U jsonargparse[signatures]
flwr~=1.3.0
numpy~=1.23.5
matplotlib~=3.7.0
#seaborn~=0.12.2
pandas~=1.5.3
scikit-learn~=1.2.1
tensorboard~=2.13.0